share the harness guards via a fluree-test-support crate
fluree-db-api and fluree-db-server carried byte-identical copies of the
reachability and env-isolation guards, roughly 250 lines each, kept in step by
a note in the module docs and nothing else. Any fix to one silently belonged in
the other.

Move the logic into fluree-test-support, a dev-dependency alongside the existing
fluree-bench-* support crates and likewise never a release artifact. Each
crate's tests/harness_coverage.rs becomes two calls passing
env!("CARGO_MANIFEST_DIR"), which expands at the call site so the crate under
test is the one checked. Adopting the guards in a third crate is now a
dev-dependency and sixteen lines rather than a copy.

The manifest parser's test moves with it and becomes a plain unit test. It
exercises a parser against a synthetic string and has nothing to do with either
crate's layout. A second unit test pins the env matcher: `std::env::set_var(`
and a bare `set_var(` match, `unset_var(` and `my_remove_var(` do not, and prose
naming the function without calling it does not either.
