+++ fluree-db-api/Cargo.toml
@@ -137,6 +137,7 @@ num-bigdecimal = { package = "bigdecimal", version = "0.4" }
 ed25519-dalek = "2.1"
 criterion = { version = "0.5", features = ["async_tokio"] }
 fluree-bench-support = { path = "../fluree-bench-support" }
+fluree-test-support = { path = "../fluree-test-support" }
 fluree-bench-alloc = { path = "../fluree-bench-alloc" }
 rand = { workspace = true }
 fluree-db-nameservice-sync = { path = "../fluree-db-nameservice-sync" }
